fix(naming): rename ArgumentMaterialType package from commands to argument
