+++ common/src/main/java/net/onelitefeather/titan/common/argument/ArgumentMaterialType.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.onelitefeather.titan.common.commands;
+package net.onelitefeather.titan.common.argument;
 
 import net.minestom.server.command.CommandSender;
 import net.minestom.server.command.builder.arguments.ArgumentString;
